+++ sdk/tables/azure-data-tables/azure/data/tables/_table_client_base.py
@@ -9,9 +9,9 @@ try:
 except ImportError:
     from urlparse import urlparse  # type: ignore
 
-from azure.data.tables._shared._error import _validate_table_name
-from azure.data.tables._base_client import parse_query
-from .._base_client import StorageAccountHostsMixin
+from ._shared._error import _validate_table_name
+from ._base_client import parse_query
+from ._base_client import StorageAccountHostsMixin
 
 
 class TableClientBase(StorageAccountHostsMixin):
