+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_base_client_async.py
@@ -21,8 +21,8 @@ from azure.core.pipeline.policies import (
 )
 from azure.core.pipeline.transport import AsyncHttpTransport
 
-from .constants import STORAGE_OAUTH_SCOPE, CONNECTION_TIMEOUT, READ_TIMEOUT
-from .authentication import SharedKeyCredentialPolicy
+from .._shared.constants import STORAGE_OAUTH_SCOPE, CONNECTION_TIMEOUT, READ_TIMEOUT
+from .._shared.authentication import SharedKeyCredentialPolicy
 from .._base_client import create_configuration
 from .._policies import (
     StorageContentValidation,
@@ -30,9 +30,9 @@ from .._policies import (
     StorageHosts,
     StorageHeadersPolicy
 )
-from ..aio._policies_async import AsyncStorageResponseHook
+from ._policies_async import AsyncStorageResponseHook
 
-from .response_handlers import process_table_error, PartialBatchErrorException
+from .._shared.response_handlers import process_table_error, PartialBatchErrorException
 
 if TYPE_CHECKING:
     from azure.core.pipeline import Pipeline
